git-svn-id: http://twcore.org/svn/trunk@1543 9f8b3893-2a26-0410-9d00-a71e40fd2f77

+++ twcore/src/twcore/bots/multibot/spaceball/Projectile.java
@@ -76,9 +76,9 @@ class Projectile {
 		return dist;
 	}
 
-	public double getXLocation() { return x + vX * getAge() / (double) 10000.0; }
+	public double getXLocation() { return x + vX * getAge() / 10000.0; }
 
-	public double getYLocation() { return y + vY * getAge() / (double) 10000.0; }
+	public double getYLocation() { return y + vY * getAge() / 10000.0; }
 
 	public int getXVelocity() { return vX; }
